Prevent image modification using rolling_ball
Subtract_background_rolling_ball (line 57) seems to modify the original image that is used as an argument (even though a different object is created for the output). That is why plotting the original image (line 67) looks like the final image (line 69-70). This issue can also be seen in the YouTube tutorial 43 (min 6:40 - 7:30).

To solve this issue, a copy of the image is created, called image_rb (line 55). The copy of the image (image_rb) is used as an argument for subtract_background_rolling_ball (line 57)